More user handling bits and pieces
Still needs a create-user form
