+++ SpiNNaker-allocserv/src/main/java/uk/ac/manchester/spinnaker/alloc/admin/UserControl.java
@@ -16,19 +16,30 @@
  */
 package uk.ac.manchester.spinnaker.alloc.admin;
 
+import static java.util.Objects.requireNonNull;
 import static uk.ac.manchester.spinnaker.alloc.DatabaseEngine.query;
 import static uk.ac.manchester.spinnaker.alloc.DatabaseEngine.transaction;
 import static uk.ac.manchester.spinnaker.alloc.DatabaseEngine.update;
 
+import java.security.Principal;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map.Entry;
+
+import javax.validation.constraints.AssertFalse;
+import javax.validation.constraints.AssertTrue;
+import javax.validation.constraints.NotBlank;
+
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationServiceException;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.InternalAuthenticationServiceException;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
@@ -262,4 +273,183 @@ public class UserControl extends SQLQueries {
 		}
 	}
 
+	/**
+	 * Describes basic information about a user that they'd use to change their
+	 * password.
+	 *
+	 * @author Donal Fellows
+	 */
+	public static class UserPassChangeModel {
+		private int userId;
+
+		private String username;
+
+		private String oldPassword;
+
+		private String newPassword;
+
+		private String newPassword2;
+
+		public UserPassChangeModel() {
+			this.userId = -1;
+			this.username = "";
+			this.oldPassword = "";
+			this.newPassword = "";
+			this.newPassword2 = "";
+		}
+
+		UserPassChangeModel(int userId, String username) {
+			this();
+			this.userId = userId;
+			this.username = requireNonNull(username);
+		}
+
+		/**
+		 * @return the user id
+		 */
+		public final int getUserId() {
+			return userId;
+		}
+
+		public void setUserId(int userId) {
+			this.userId = userId;
+		}
+
+		/**
+		 * @return the username
+		 */
+		public final String getUsername() {
+			return username;
+		}
+
+		public void setUsername(String username) {
+			this.username = username;
+		}
+
+		/**
+		 * @return the old password
+		 */
+		@NotBlank(message = "old password must be supplied")
+		public String getOldPassword() {
+			return oldPassword;
+		}
+
+		public void setOldPassword(String password) {
+			this.oldPassword = password;
+		}
+
+		/**
+		 * @return the first copy of the new password
+		 */
+		@NotBlank(message = "new password must be supplied")
+		public String getNewPassword() {
+			return newPassword;
+		}
+
+		public void setNewPassword(String newPassword) {
+			this.newPassword = newPassword;
+		}
+
+		/**
+		 * @return the second copy of the new password
+		 */
+		@NotBlank(message = "second copy of new password must be supplied")
+		public String getNewPassword2() {
+			return newPassword2;
+		}
+
+		public void setNewPassword2(String newPassword2) {
+			this.newPassword2 = newPassword2;
+		}
+
+		@AssertFalse(message = "old and new passwords must be different")
+		boolean isNewPasswordSameAsOld() {
+			return newPassword.equals(oldPassword);
+		}
+
+		@AssertTrue(
+				message = "second copy of new password must be same as first")
+		boolean isNewPasswordMatched() {
+			return newPassword.equals(newPassword2);
+		}
+	}
+
+	/**
+	 * Get a model for updating the local password of the current user.
+	 *
+	 * @param principal
+	 *            The current user
+	 * @return User model object. Password fields are unfilled.
+	 * @throws AuthenticationException
+	 *             If the user cannot change their password here for some
+	 *             reason.
+	 * @throws SQLException
+	 *             If DB access fails.
+	 */
+	public UserPassChangeModel getUserForPrincipal(Principal principal)
+			throws AuthenticationException, SQLException {
+		try (Connection c = db.getConnection();
+				Query q = query(c, "SELECT user_id, user_name FROM user_info "
+						+ "WHERE user_name = :username "
+						+ "AND encrypted_password IS NOT NULL LIMIT 1")) {
+			Row row = q.call1(principal.getName()).orElseThrow(
+					// OpenID-authenticated user; go away
+					() -> new AuthenticationServiceException(
+							"user is managed externally; "
+									+ "cannot manage password here"));
+			return new UserPassChangeModel(row.getInt("user_id"),
+					row.getString("user_name"));
+		}
+	}
+
+	/**
+	 * Update the local password of the current user based on a filled out model
+	 * previously provided.
+	 *
+	 * @param principal
+	 *            The current user
+	 * @param user
+	 *            Valid user model object with password fields filled.
+	 * @return Replacement user model object. Password fields are unfilled.
+	 * @throws AuthenticationException
+	 *             If the user cannot change their password here for some
+	 *             reason.
+	 * @throws SQLException
+	 *             If DB access fails.
+	 */
+	public UserPassChangeModel updateUserOfPrincipal(Principal principal,
+			UserPassChangeModel user)
+			throws AuthenticationException, SQLException {
+		try (Connection c = db.getConnection();
+				Query getPassword = query(c,
+						"SELECT user_id, user_name, encrypted_password "
+								+ "FROM user_info "
+								+ "WHERE user_name = :user_name "
+								+ "AND encrypted_password IS NOT NULL LIMIT 1");
+				Update setPassword = update(c, SET_USER_PASS)) {
+			return transaction(c, () -> {
+				Row row = getPassword.call1(principal.getName()).orElseThrow(
+						// OpenID-authenticated user; go away
+						() -> new AuthenticationServiceException(
+								"user is managed externally; "
+										+ "cannot change password here"));
+				UserPassChangeModel baseUser = new UserPassChangeModel(
+						row.getInt("user_id"), row.getString("user_name"));
+				if (!passwordEncoder.matches(user.oldPassword,
+						row.getString("encrypted_password"))) {
+					throw new BadCredentialsException("bad password");
+				}
+				// Validate change; this should never fail but...
+				if (!user.newPassword.equals(user.newPassword2)) {
+					throw new BadCredentialsException("bad password");
+				}
+				if (setPassword.call(passwordEncoder.encode(user.newPassword),
+						baseUser.userId) != 1) {
+					throw new InternalAuthenticationServiceException(
+							"failed to update database");
+				}
+				return baseUser;
+			});
+		}
+	}
 }
